requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.2

# Database
asyncpg>=0.29.0
qdrant-client>=1.6.0

# Configuration & Environment
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Cache & Storage
redis>=5.0.1

# System Utilities
psutil>=5.9.0

# CLI & UI
click>=8.1.7
rich>=13.7.0

# WebSocket support for MCP
websockets>=12.0